Cleaned up Console implementation
Disentangled the ezConsole and ezQuakeConsole implementation.
Moved ezQuakeConsole and ezCommandInterpreterLua into GameEngine.
Fixed that switching the mouse cursor visibility at runtime didn't work.
Turns out calling ShowCursor from another thread has no effect.
Fixed that RenderFps and RenderConsole where called on another thread than the rest of the rendering.
