+++ Code/Engine/Core/Console/Implementation/Console.cpp
@@ -1,217 +1,10 @@
 #include <Core/CorePCH.h>
 
-#include <Core/Console/LuaInterpreter.h>
-#include <Core/Console/QuakeConsole.h>
+#include <Core/Console/Console.h>
 #include <Foundation/IO/FileSystem/FileReader.h>
 #include <Foundation/IO/FileSystem/FileWriter.h>
 
 EZ_ENUMERABLE_CLASS_IMPLEMENTATION(ezConsoleFunctionBase);
-
-ezQuakeConsole::ezQuakeConsole()
-{
-  ClearInputLine();
-
-  m_bLogOutputEnabled = false;
-  m_bDefaultInputHandlingInitialized = false;
-  m_uiMaxConsoleStrings = 1000;
-
-  EnableLogOutput(true);
-
-#ifdef BUILDSYSTEM_ENABLE_LUA_SUPPORT
-  SetCommandInterpreter(EZ_DEFAULT_NEW(ezCommandInterpreterLua));
-#endif
-}
-
-ezQuakeConsole::~ezQuakeConsole()
-{
-  EnableLogOutput(false);
-}
-
-void ezQuakeConsole::AddConsoleString(ezStringView sText, ezConsoleString::Type type)
-{
-  EZ_LOCK(m_Mutex);
-
-  m_ConsoleStrings.PushFront();
-
-  ezConsoleString& cs = m_ConsoleStrings.PeekFront();
-  cs.m_sText = sText;
-  cs.m_Type = type;
-
-  if (m_ConsoleStrings.GetCount() > m_uiMaxConsoleStrings)
-    m_ConsoleStrings.PopBack(m_ConsoleStrings.GetCount() - m_uiMaxConsoleStrings);
-
-  ezConsole::AddConsoleString(sText, type);
-}
-
-const ezDeque<ezConsoleString>& ezQuakeConsole::GetConsoleStrings() const
-{
-  if (m_bUseFilteredStrings)
-  {
-    return m_FilteredConsoleStrings;
-  }
-
-  return m_ConsoleStrings;
-}
-
-void ezQuakeConsole::LogHandler(const ezLoggingEventData& data)
-{
-  ezConsoleString::Type type = ezConsoleString::Type::Default;
-
-  switch (data.m_EventType)
-  {
-    case ezLogMsgType::GlobalDefault:
-    case ezLogMsgType::Flush:
-    case ezLogMsgType::BeginGroup:
-    case ezLogMsgType::EndGroup:
-    case ezLogMsgType::None:
-    case ezLogMsgType::ENUM_COUNT:
-    case ezLogMsgType::All:
-      return;
-
-    case ezLogMsgType::ErrorMsg:
-      type = ezConsoleString::Type::Error;
-      break;
-
-    case ezLogMsgType::SeriousWarningMsg:
-      type = ezConsoleString::Type::SeriousWarning;
-      break;
-
-    case ezLogMsgType::WarningMsg:
-      type = ezConsoleString::Type::Warning;
-      break;
-
-    case ezLogMsgType::SuccessMsg:
-      type = ezConsoleString::Type::Success;
-      break;
-
-    case ezLogMsgType::InfoMsg:
-      break;
-
-    case ezLogMsgType::DevMsg:
-      type = ezConsoleString::Type::Dev;
-      break;
-
-    case ezLogMsgType::DebugMsg:
-      type = ezConsoleString::Type::Debug;
-      break;
-  }
-
-  ezStringBuilder sFormat;
-  sFormat.SetPrintf("%*s", data.m_uiIndentation, "");
-  sFormat.Append(data.m_sText);
-
-  AddConsoleString(sFormat.GetData(), type);
-}
-
-void ezQuakeConsole::InputStringChanged()
-{
-  m_bUseFilteredStrings = false;
-  m_FilteredConsoleStrings.Clear();
-
-  if (m_sInputLine.StartsWith("*"))
-  {
-    ezStringBuilder input = m_sInputLine;
-
-    input.Shrink(1, 0);
-    input.Trim(" ");
-
-    if (input.IsEmpty())
-      return;
-
-    m_FilteredConsoleStrings.Clear();
-    m_bUseFilteredStrings = true;
-
-    for (const auto& e : m_ConsoleStrings)
-    {
-      if (e.m_sText.FindSubString_NoCase(input))
-      {
-        m_FilteredConsoleStrings.PushBack(e);
-      }
-    }
-
-    Scroll(0); // clamp scroll position
-  }
-}
-
-void ezQuakeConsole::EnableLogOutput(bool bEnable)
-{
-  if (m_bLogOutputEnabled == bEnable)
-    return;
-
-  m_bLogOutputEnabled = bEnable;
-
-  if (bEnable)
-  {
-    ezGlobalLog::AddLogWriter(ezMakeDelegate(&ezQuakeConsole::LogHandler, this));
-  }
-  else
-  {
-    ezGlobalLog::RemoveLogWriter(ezMakeDelegate(&ezQuakeConsole::LogHandler, this));
-  }
-}
-
-void ezQuakeConsole::SaveState(ezStreamWriter& inout_stream) const
-{
-  EZ_LOCK(m_Mutex);
-
-  const ezUInt8 uiVersion = 1;
-  inout_stream << uiVersion;
-
-  inout_stream << m_InputHistory.GetCount();
-  for (ezUInt32 i = 0; i < m_InputHistory.GetCount(); ++i)
-  {
-    inout_stream << m_InputHistory[i];
-  }
-
-  inout_stream << m_BoundKeys.GetCount();
-  for (auto it = m_BoundKeys.GetIterator(); it.IsValid(); ++it)
-  {
-    inout_stream << it.Key();
-    inout_stream << it.Value();
-  }
-}
-
-void ezQuakeConsole::LoadState(ezStreamReader& inout_stream)
-{
-  EZ_LOCK(m_Mutex);
-
-  ezUInt8 uiVersion = 0;
-  inout_stream >> uiVersion;
-
-  if (uiVersion == 1)
-  {
-    ezUInt32 count = 0;
-    inout_stream >> count;
-    m_InputHistory.SetCount(count);
-
-    for (ezUInt32 i = 0; i < m_InputHistory.GetCount(); ++i)
-    {
-      inout_stream >> m_InputHistory[i];
-    }
-
-    inout_stream >> count;
-
-    ezString sKey;
-    ezString sValue;
-
-    for (ezUInt32 i = 0; i < count; ++i)
-    {
-      inout_stream >> sKey;
-      inout_stream >> sValue;
-
-      m_BoundKeys[sKey] = sValue;
-    }
-  }
-}
-
-void ezCommandInterpreterState::AddOutputLine(const ezFormatString& text, ezConsoleString::Type type /*= ezCommandOutputLine::Type::Default*/)
-{
-  auto& line = m_sOutput.ExpandAndGetRef();
-  line.m_Type = type;
-
-  ezStringBuilder tmp;
-  line.m_sText = text.GetText(tmp);
-}
 
 ezColor ezConsoleString::GetColor() const
 {
@@ -255,11 +48,6 @@ ezColor ezConsoleString::GetColor() const
 
   return ezColor::White;
 }
-
-//////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////
 
 ezConsole::ezConsole() = default;
